+++ buildme.sh
@@ -24,9 +24,9 @@ if [ "$build_status" == 0 ]; then
   echo "Docker build succeed"
   rm -rfv dive.log||true
   rm -rfv ./.coverage.*||true
-  trivy --output .coverage."$version"_trivy.txt "$release":"$version"
+  trivy image --output .coverage."$version"_trivy.txt "$release":"$version"
   dive --ci "$release":"$version" > .coverage."$version"_dive.txt
-  dockle -f json -o .coverage."$version"_dockle.txt "$release":"$version"
+  sudo dockle -f json -o .coverage."$version"_dockle.txt "$release":"$version"
 else
  echo "Docker build failed, exiting now"
 fi